requirements.txt
streamlit 
sentence-transformers 
scikit-learn
faiss-cpu
numpy